Fix: Move building of postfix-exporter to a separate action and fix build
To not repeat outselves, building of postfix-exporter was moved to a
separate action.

Helm now always references the lastest version, as there's no sense in
having specific versions which are all the same.

Lastly, the build fails in the test phase, as it relies on an external
dependency and hence tests have been (temporarily, at least) removed
from the build cycle. This should prove to be a large issue, as we're
building from a fixed SHA hash which we know to be valid.