Fix for #174: Incrase healthcheck timeout
Default timeout for healthcheck by kubernetes is too low. As we wait up
for 2 seconds for the Postfix and DKIM to respond, this exceeds the
threshold and the healthcheck fails with:

```
Readiness probe failed: command "sh -c /scripts/healthcheck.sh" timed out
```

By simply increasing `timeoutSeconds` this issue is resolved.