Fix for #192: Automatically add domains to provided usernames for SASL
So, according to the documentation, usernames must always include a
domain for SASL.

In other words. User cannot be `johhny` but `johhny@example.org`.
Further info can be found on this ticket: https://github.com/bokysan/docker-postfix/issues/192

This commit will automatically append domain if one is not provided in
`SMTPD_SASL_USERS`.